+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/states/disconnected_state.rs
@@ -66,11 +66,6 @@ impl TunnelStateHandler for DisconnectedState {
                         NextTunnelState::NewState(ConnectingState::enter(0, None, shared_state).await)
                     },
                     TunnelCommand::Disconnect => NextTunnelState::SameState(self),
-                    TunnelCommand::SetAllowLan(allow_lan, complete_tx) => {
-                        _ = shared_state.set_allow_lan(allow_lan);
-                        _ = complete_tx.send(());
-                        NextTunnelState::SameState(self)
-                    },
                     TunnelCommand::SetTunnelSettings(tunnel_settings) => {
                         shared_state.tunnel_settings = tunnel_settings;
                         NextTunnelState::SameState(self)