+++ nym-vpn-core/crates/nym-vpnd/src/service/vpn_service.rs
@@ -1,10 +1,9 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use std::{path::PathBuf, pin::Pin};
-
 use bip39::Mnemonic;
 use futures::{FutureExt, StreamExt, future::Fuse, pin_mut};
+use std::{net::IpAddr, path::PathBuf, pin::Pin};
 use time::{OffsetDateTime, format_description::well_known::Rfc3339};
 use tokio::{
     sync::{broadcast, mpsc, oneshot, watch},
@@ -66,6 +65,7 @@ pub enum VpnServiceCommand {
     SetAllowLan(oneshot::Sender<()>, bool),
     SetEnableBridges(oneshot::Sender<()>, bool),
     SetResidentialExit(oneshot::Sender<()>, bool),
+    SetCustomDns(oneshot::Sender<()>, Option<Vec<IpAddr>>),
     SetNetwork(oneshot::Sender<Result<(), SetNetworkError>>, String),
     GetSystemMessages(oneshot::Sender<Vec<SystemMessage>>, ()),
     GetNetworkCompatibility(oneshot::Sender<Option<NetworkCompatibility>>, ()),
@@ -726,7 +726,8 @@ impl NymVpnService {
                 let _ = tx.send(());
             }
             VpnServiceCommand::SetAllowLan(tx, allow_lan) => {
-                self.handle_set_allow_lan(allow_lan, tx).await;
+                self.handle_set_allow_lan(allow_lan).await;
+                let _ = tx.send(());
             }
             VpnServiceCommand::SetEnableBridges(tx, enable_bridges) => {
                 self.handle_set_enable_bridges(enable_bridges).await;
@@ -734,6 +735,10 @@ impl NymVpnService {
             }
             VpnServiceCommand::SetResidentialExit(tx, residential_exit) => {
                 self.handle_set_residential_exit(residential_exit).await;
+                let _ = tx.send(());
+            }
+            VpnServiceCommand::SetCustomDns(tx, custom_dns) => {
+                self.handle_set_custom_dns(custom_dns).await;
                 let _ = tx.send(());
             }
             VpnServiceCommand::SetNetwork(tx, network) => {
@@ -892,11 +897,9 @@ impl NymVpnService {
         self.update_tunnel_settings_with_throttle();
     }
 
-    async fn handle_set_allow_lan(&mut self, allow_lan: bool, complete_tx: oneshot::Sender<()>) {
+    async fn handle_set_allow_lan(&mut self, allow_lan: bool) {
         self.config_manager.set_allow_lan(allow_lan).await;
-        _ = self
-            .command_sender
-            .send(TunnelCommand::SetAllowLan(allow_lan, complete_tx));
+        self.update_tunnel_settings_with_throttle();
     }
 
     async fn handle_set_enable_bridges(&mut self, enable_bridges: bool) {
@@ -908,6 +911,11 @@ impl NymVpnService {
         self.config_manager
             .set_residential_exit(residential_exit)
             .await;
+        self.update_tunnel_settings_with_throttle();
+    }
+
+    async fn handle_set_custom_dns(&mut self, custom_dns: Option<Vec<IpAddr>>) {
+        self.config_manager.set_custom_dns(custom_dns).await;
         self.update_tunnel_settings_with_throttle();
     }
 
@@ -1027,6 +1035,7 @@ impl NymVpnService {
 
         let entry_point = entry.unwrap_or(self.config_manager.config().entry_point.clone());
         let exit_point = exit.unwrap_or(self.config_manager.config().exit_point.clone());
+        let custom_dns = options.dns.as_ref().map(|ip_addr| vec![*ip_addr]);
         let config = VpnServiceConfig {
             entry_point,
             exit_point,
@@ -1034,7 +1043,6 @@ impl NymVpnService {
             enable_two_hop: options.enable_two_hop,
             enable_bridges: options.enable_bridges,
             netstack: options.netstack,
-            dns: options.dns,
             allow_lan: true, // always true to support legacy behavior
             min_mixnode_performance: None,
             min_gateway_mixnet_performance: None,
@@ -1042,6 +1050,7 @@ impl NymVpnService {
             disable_poisson_rate: options.disable_poisson_rate,
             disable_background_cover_traffic: options.disable_background_cover_traffic,
             residential_exit: false,
+            custom_dns,
         };
 
         self.config_manager.set_config(config).await;